// my-app/src/modules/manage/order/models/OrderInfResponse.ts
export type OrderInfResponse = {
    id: string;
    orderHdrID: string | null;
    originalID: string | null;
    itemCode: string | null;
    shipmentId: string | null;
    originalItemCode: string | null;
    status: string;
    saleOrderCode: string | null;
    senderCode: string | null;
    senderContractNumber: string | null;
    senderName: string | null;
    senderPhone: string | null;
    senderEmail: string | null;
    senderAddress: string | null;
    senderProvinceCode: string | null;
    senderProvinceName: string | null;
    senderDistrictCode: string | null;
    senderDistrictName: string | null;
    senderCommuneCode: string | null;
    senderCommuneName: string | null;
    senderPostcode: string | null;
    receiverCode: string | null;
    receiverName: string | null;
    receiverContractNumber: string | null;
    receiverAddress: string | null;
    receiverProvinceCode: string | null;
    receiverProvinceName: string | null;
    receiverDistrictCode: string | null;
    receiverDistrictName: string | null;
    receiverCommCode: string | null;
    receiverCommName: string | null;
    receiverPostcode: string | null;
    receiverPhone: string | null;
    receiverEmail: string | null;
    serviceCode: string | null;
    totalFee: number | null;
    mainFee: number | null;
    mainFeeBeforeTax: number | null;
    mainTax: number | null;
    vasFee: number | null;
    codAmount: number | null;
    addonService: string | null;
    additionRequest: string | null;
    weight: number | null;
    length: number | null;
    width: number | null;
    height: number | null;
    dimWeight: number | null;
    priceWeight: number | null;
    deliveryInstruction: string | null;
    quantity: number | null;
    contentNote: string | null;
    contentDetail: string | null;
    sendType: string | null;
    isBroken: boolean | null;
    deliveryTime: string | null;
    deliveryRequire: string | null;
    vehicle: string | null;
    awbNumber: string | null;
    voucher: string | null;
    orgCodeCollect: string | null;
    orgCodeAccept: string | null;
    createdBy: string | null;
    createdDate: string | null;
    updatedBy: string | null;
    updatedDate: string | null;
    acceptedDate: string | null;
    sortingCode: string | null;
    paymentStatus: string | null;
    caseTypeName: string | null;
    caseStatus: string | null;
    source: string | null;
    inputType: string | null;
    inputMethod: string | null;
    documents: string | null;
    packageInfo: string | null;
};